+++ noodles-util/src/variant/async/io.rs
@@ -1,5 +1,6 @@
 //! Async variant format I/O.
 
 pub mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};